Changed heading levels to match document structure

+++ system/application/views/news/news.php
@@ -3,38 +3,42 @@ function printarticlelink($article, $blurb = false) {
 	echo('	<div class="Entry">'."\n");
 	echo('		<a href="/news/'.$article['article_type'].'/'.$article['id'].'">'."\n");
 	echo('			<img src="'.$article['photo_url'].'" alt="'.$article['photo_title'].'" title="'.$article['photo_title'].'" />'."\n");
-	echo('			<h5>'.$article['heading'].'</h5>'."\n");
 	echo('		</a>'."\n");
+	echo('		<h3>'."\n");
+	echo('			<a href="/news/'.$article['article_type'].'/'.$article['id'].'">'."\n");
+	echo('				'.$article['heading']."\n");
+	echo('			</a>'."\n");
+	echo('		</h3>'."\n");
 	foreach($article['authors'] as $reporter)
 		echo('		<p><a href="/contact">'.$reporter['name'].'</a></p>'."\n");
 	echo('		<p>'.$article['date'].'</p>'."\n");
 	echo('		'.anchor('/news/'.$article['article_type'].'/'.$article['id'], 'Read more...')."\n");
-	if ($blurb)
+	if (array_key_exists('blurb', $article) && $article['blurb'] != '')
 		echo('		<p>'.$article['blurb'].'</p>'."\n");
 	echo('	</div>'."\n");
 }
 ?>
 
 <div id="RightColumn">
-	<h4 class="First"><?php echo($latest_heading); ?></h4>
+	<h2 class="first"><?php echo($latest_heading); ?></h2>
 <?php
 foreach($news_previews as $preview)
 	printarticlelink($preview, true);
 ?>
 	
-	<h4><?php echo($other_heading); ?></h4>
+	<h2><?php echo($other_heading); ?></h2>
 <?php
 foreach ($news_others as $other)
 	printarticlelink($other, false);
 
 if (count($main_article['related_articles']) > 0)
-	echo('	<h4>'.$related_heading.'</h4>'."\n");
+	echo('	<h2>'.$related_heading.'</h2>'."\n");
 
 foreach ($main_article['related_articles'] as $related)
 	printarticlelink($related, false);
 
 foreach($main_article['fact_boxes'] as $fact_box) {
-	echo('	<h4>'.$fact_box['title'].'</h4>'."\n");
+	echo('	<h2>'.$fact_box['title'].'</h2>'."\n");
 	echo('	<div class="Entry">'."\n");
 	echo('		'.$fact_box['wikitext']."\n");
 	echo('	</div>'."\n");
@@ -42,13 +46,11 @@ foreach($main_article['fact_boxes'] as $fact_box) {
 ?>
 </div>
 
-
-
-
-	<div class='grey_box ArticleColumn'>
-		<h1><?php echo $main_article['heading']; ?></h1>
+<div id="MainColumn">
+	<div class="BlueBox">
+		<h2><?php echo $main_article['heading']; ?></h2>
 		<?php if ($main_article['subheading'] != '') { ?>
-			<h2><?php echo $main_article['subheading']; ?></h2>
+			<h3><?php echo $main_article['subheading']; ?></h3>
 		<?php } ?>
 		<?php if ($main_article['subtext'] != '') { ?>
 		<div class='intro'><?php echo $main_article['subtext']; ?></div>
@@ -56,20 +58,20 @@ foreach($main_article['fact_boxes'] as $fact_box) {
 
 		<?php $this->byline->load(); ?>
 
-        <p><?php echo $main_article['text']; ?></p>
+        <?php echo $main_article['text']; ?>
 
 		<?php if (isset($office_preview)) { ?>
 			<p class='form'><button class="button" onclick="window.location='/office/news/<?php echo $main_article['id']; ?>';">GO BACK TO NEWS OFFICE</button></p>
 		<?php } ?>
-		<br style='clear: both;' />
 	</div>
 	<?php if (count($main_article['links']) > 0) { ?>
-		<div class='blue_box'>
-			<h2><?php echo $links_heading; ?></h2>
-			<ul>
-			<?php foreach ($main_article['links'] as $link) {
-				echo '<li><a href=\'' . $link['url'] . '\' target=\'_blank\'>' . $link['name'] . '</a></li>';
-			} ?>
-			</ul>
-		</div>
+	<div class="BlueBox">
+		<h2><?php echo $links_heading; ?></h2>
+		<ul>
+		<?php foreach ($main_article['links'] as $link) {
+			echo '<li><a href=\'' . $link['url'] . '\' target=\'_blank\'>' . $link['name'] . '</a></li>';
+		} ?>
+		</ul>
+	</div>
 	<?php } ?>
+</div>
